Expand Postgres file-node visibility coverage
Move large file-node test helpers out of the entry module so the test
surface stays below the module-size limit while keeping existing re-exports
stable. Run the revocation and group-membership visibility scenarios against
Postgres as well as SQLite through the shared harness.

Document the migration watchdog timeout contract more precisely and log
whether cancelled migration work finishes within the grace window before the
watchdog returns its timeout error.
